Issue #820 Get rid of SerializerConfiguration
* Moved Type enum to DefaultSerializerConfiguration
