+++ impl/src/main/java/org/ehcache/impl/config/persistence/CacheManagerPersistenceConfiguration.java
@@ -20,7 +20,6 @@ import org.ehcache.CacheManager;
 import org.ehcache.PersistentCacheManager;
 import org.ehcache.config.builders.CacheManagerConfiguration;
 import org.ehcache.config.builders.CacheManagerBuilder;
-import org.ehcache.core.config.persistence.PersistenceConfiguration;
 import org.ehcache.core.spi.service.LocalPersistenceService;
 
 import java.io.File;
