+++ impl/src/main/java/org/ehcache/impl/config/serializer/DefaultSerializerConfiguration.java
@@ -16,15 +16,15 @@
 
 package org.ehcache.impl.config.serializer;
 
-import org.ehcache.core.config.serializer.SerializerConfiguration;
 import org.ehcache.impl.internal.classes.ClassInstanceConfiguration;
 import org.ehcache.spi.serialization.SerializationProvider;
 import org.ehcache.spi.serialization.Serializer;
+import org.ehcache.spi.service.ServiceConfiguration;
 
 /**
  * @author Ludovic Orban
  */
-public class DefaultSerializerConfiguration<T> extends ClassInstanceConfiguration<Serializer<T>> implements SerializerConfiguration<SerializationProvider> {
+public class DefaultSerializerConfiguration<T> extends ClassInstanceConfiguration<Serializer<T>> implements ServiceConfiguration<SerializationProvider> {
 
   private final Type type;
 
@@ -43,8 +43,20 @@ public class DefaultSerializerConfiguration<T> extends ClassInstanceConfiguratio
     return SerializationProvider.class;
   }
 
+  /**
+   * Get the type of the serializer configured
+   *
+   * @return the type
+   */
   public Type getType() {
     return type;
   }
 
+  /**
+   * Serialization provider types
+   */
+  public enum Type {
+    KEY,
+    VALUE,
+  }
 }
