+++ impl/src/main/java/org/ehcache/impl/internal/persistence/DefaultLocalPersistenceServiceFactory.java
@@ -16,7 +16,7 @@
 
 package org.ehcache.impl.internal.persistence;
 
-import org.ehcache.core.config.persistence.PersistenceConfiguration;
+import org.ehcache.impl.config.persistence.PersistenceConfiguration;
 import org.ehcache.core.spi.service.LocalPersistenceService;
 import org.ehcache.impl.persistence.DefaultLocalPersistenceService;
 import org.ehcache.spi.service.ServiceCreationConfiguration;
