+++ impl/src/test/java/org/ehcache/impl/persistence/DefaultLocalPersistenceServiceTest.java
@@ -16,8 +16,7 @@
 
 package org.ehcache.impl.persistence;
 
-import org.ehcache.core.config.persistence.DefaultPersistenceConfiguration;
-import org.ehcache.impl.persistence.DefaultLocalPersistenceService;
+import org.ehcache.impl.config.persistence.DefaultPersistenceConfiguration;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
